Move tests for the disk selection module to a new file
Move all tests for the disk selection module to a new file.
