+++ tests/nosetests/pyanaconda_tests/module_disk_select_test.py
@@ -1,0 +1,63 @@
+#
+# Copyright (C) 2019  Red Hat, Inc.
+#
+# This copyrighted material is made available to anyone wishing to use,
+# modify, copy, or redistribute it subject to the terms and conditions of
+# the GNU General Public License v.2, or (at your option) any later version.
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY expressed or implied, including the implied warranties of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General
+# Public License for more details.  You should have received a copy of the
+# GNU General Public License along with this program; if not, write to the
+# Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
+# 02110-1301, USA.  Any Red Hat trademarks that are incorporated in the
+# source code or documentation are not subject to the GNU General Public
+# License and may only be used or replicated with the express permission of
+# Red Hat, Inc.
+#
+# Red Hat Author(s): Vendula Poncova <vponcova@redhat.com>
+#
+import unittest
+
+from pyanaconda.modules.common.constants.objects import DISK_SELECTION
+from pyanaconda.modules.storage.disk_selection import DiskSelectionModule
+from pyanaconda.modules.storage.disk_selection.selection_interface import DiskSelectionInterface
+from tests.nosetests.pyanaconda_tests import check_dbus_property
+
+
+class DiskSelectionInterfaceTestCase(unittest.TestCase):
+    """Test DBus interface of the disk selection module."""
+
+    def setUp(self):
+        """Set up the module."""
+        self.disk_selection_module = DiskSelectionModule()
+        self.disk_selection_interface = DiskSelectionInterface(self.disk_selection_module)
+
+    def _test_dbus_property(self, *args, **kwargs):
+        check_dbus_property(
+            self,
+            DISK_SELECTION,
+            self.disk_selection_interface,
+            *args, **kwargs
+        )
+
+    def selected_disks_property_test(self):
+        """Test the selected disks property."""
+        self._test_dbus_property(
+            "SelectedDisks",
+            ["sda", "sdb"]
+        )
+
+    def ignored_disks_property_test(self):
+        """Test the ignored disks property."""
+        self._test_dbus_property(
+            "IgnoredDisks",
+            ["sda", "sdb"]
+        )
+
+    def protected_disks_property_test(self):
+        """Test the protected disks property."""
+        self._test_dbus_property(
+            "ProtectedDevices",
+            ["sda", "sdb"]
+        )
